% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alignment.R
\name{align_reads}
\alias{align_reads}
\title{Align Reads}
\usage{
align_reads(go_obj, outdir, cores = 1)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{Directory to save aligned files to}

\item{cores}{Number of CPU cores/threads to use}
}
\description{
Align reads to genome using STAR
}
