% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_count.R
\name{export_counts}
\alias{export_counts}
\title{Export Feature Counts}
\usage{
export_counts(go_obj, outdir)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{Output directory for table}
}
\value{
gostripes object and tab separated table of feature counts.
}
\description{
Export feature counts as a table
}
\examples{
R1_fastq <- system.file("extdata", "S288C_R1.fastq", package = "gostripes")
R2_fastq <- system.file("extdata", "S288C_R2.fastq", package = "gostripes")
rRNA <- system.file("extdata", "Sc_rRNA.fasta", package = "gostripes")
assembly <- system.file("extdata", "Saccharomyces_cerevisiae.R64-1-1.dna_sm.toplevel.fa", package = "gostripes")
annotation <- system.file("extdata", "Saccharomyces_cerevisiae.R64-1-1.99.gtf", package = "gostripes")

sample_sheet <- tibble::tibble(
  "sample_name" = "stripeseq", "replicate_ID" = 1,
  "R1_read" = R1_fastq, "R2_read" = R2_fastq
)

go_object <- gostripes(sample_sheet) \%>\%
  process_reads("./scratch/cleaned_fastq", rRNA) \%>\%
  fastq_quality("./scratch/fastqc_reports") \%>\%
  genome_index(assembly, annotation, "./scratch/genome_index") \%>\%
  align_reads("./scratch/aligned") \%>\%
  process_bams("./scratch/cleaned_bams") \%>\%
  count_features(annotation) \%>\%
  export_counts("./scratch/counts")

}
