% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_quality.R
\name{fastq_quality}
\alias{fastq_quality}
\title{Fastq Quality}
\usage{
fastq_quality(go_obj, outdir, fastq_type = "both", cores = 1)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{Output directory for FastQC reports}

\item{fastq_type}{Either 'raw', 'processed', or 'both'}

\item{cores}{Number of CPU cores}
}
\description{
Generate quality reports for FASTQ files
}
