% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_processing.R
\name{filter_flags}
\alias{filter_flags}
\title{Filter Flags}
\usage{
filter_flags(bam_file, sample_name, outdir, cores)
}
\arguments{
\item{bam_file}{Aligned bam file to process}

\item{sample_name}{Name of sample}

\item{outdir}{Output directory}

\item{cores}{Number of CPU cores available}
}
\description{
Remove PCR duplicates and other undesirable reads
}
