% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_processing.R
\name{filter_soft}
\alias{filter_soft}
\title{Filter Soft-clipped}
\usage{
filter_soft(deduped_bam, sample_name, outdir, seq_mode)
}
\arguments{
\item{sample_name}{Name of the sample}

\item{outdir}{Output directory}

\item{seq_mode}{Whether the bam is paired-end or single-end}

\item{flagfiltered_bam}{Bam file with PCR duplicates and undesirable flags filtered out}
}
\description{
Filter reads with too many soft-clipped bases
}
