% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_processing.R
\name{process_reads}
\alias{process_reads}
\title{Process Reads}
\usage{
process_reads(go_obj, outdir, contamination_fasta, cores = 1)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{output directory for filtered and trimmed reads}

\item{contamination_fasta}{fasta file containing contaminants to remove, such as rRNA}

\item{cores}{Number of CPU core/threads to use}
}
\description{
Remove rRNA contamination, low complexity reads, and trim UMI-Spacer-GGG
}
