% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_processing.R
\name{read_structure}
\alias{read_structure}
\title{Check R1 Structure}
\usage{
read_structure(
  R1_read,
  R2_read,
  sample_name,
  outdir,
  structure_regex = "^[ATGCN]{8}TATAGGG"
)
}
\arguments{
\item{R1_read}{File name and directory for R1 read}

\item{R2_read}{File name and directory for R2 read}

\item{sample_name}{Name of the sample}

\item{outdir}{Output directory for the properly structured reads}

\item{structure_regex}{Regular expression for structure of read 5' end}
}
\description{
Ensure proper structure of R1 read, and remove spacer and ribo-Gs
}
