% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_processing.R
\name{remove_contaminants}
\alias{remove_contaminants}
\title{Remove Contaminants}
\usage{
remove_contaminants(
  proper_R1,
  proper_R2,
  sample_name,
  contamination_fasta,
  outdir,
  cores
)
}
\arguments{
\item{proper_R1}{R1 reads with proper structure}

\item{proper_R2}{R2 reads where R1 had proper structure}

\item{cores}{Number of CPU cores available}
}
\description{
Remove low complexity reads and contaminant reads such as rRNA
}
