% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_processing.R
\name{remove_extra}
\alias{remove_extra}
\title{Remove Extra}
\usage{
remove_extra(
  stashed_R1,
  stashed_R2,
  sample_name,
  outdir,
  extra_bases = "TATAGGG"
)
}
\arguments{
\item{stashed_R1}{R1 read with UMI stashed in read name}

\item{stashed_R2}{R2 read with UMI stashed in read name}

\item{sample_name}{Name of the sample}

\item{outdir}{Output directory of reads with spacer adn ribo-Gs trimmed}

\item{extra_bases}{The structure of the extra 5' bases}
}
\description{
Remove spacer (TATA) and ribo-Gs (GGG) from 5' end of read
}
