% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_processing.R
\name{stash_umi}
\alias{stash_umi}
\title{Stash UMI}
\usage{
stash_umi(
  deconned_R1,
  deconned_R2,
  sample_name,
  outdir,
  umi_pattern = "NNNNNNNN"
)
}
\arguments{
\item{deconned_R1}{R1 read with contaminants removed}

\item{deconned_R2}{R2 read with contaminants removed}

\item{outdir}{Output directory}

\item{umi_pattern}{UMI pattern on 5' end of read}
}
\description{
Add UMI to read name to facilitate PCR duplicate removal
}
