% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_processing.R
\name{trim_reads}
\alias{trim_reads}
\title{Trim R1 Read}
\usage{
trim_reads(go_obj, outdir)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{output directory for trimmed reads}
}
\description{
Check for proper structure of R1 read and then trim it
}
