% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_analysis.R
\name{call_TSRs}
\alias{call_TSRs}
\title{Call TSRs}
\usage{
call_TSRs(go_obj, outdir)
}
\arguments{
\item{go_obj}{gostripes object}

\item{threshold}{TSSs with read count below threshold will be discarded}

\item{clust_dist}{TSSs within this number of base pairs will be clustered}
}
\description{
Basic TSS clustering into TSRs based on naive global read threshold
}
