% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_count.R
\name{count_features}
\alias{count_features}
\title{Count Features}
\usage{
count_features(go_obj, genome_annotation, cores = 1)
}
\arguments{
\item{go_obj}{gostripes object}

\item{genome_annotation}{Genome annotation in GTF file format}

\item{cores}{Number of CPU cores available}
}
\description{
Count reads associated with annotated features
}
