% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_count.R
\name{export_counts}
\alias{export_counts}
\title{Export Feature Counts}
\usage{
export_counts(go_obj, outdir)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{Output directory for table}
}
\description{
Export feature counts as a table
}
