% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_analysis.R
\name{export_TSRs}
\alias{export_TSRs}
\title{Export TSRs}
\usage{
export_TSRs(go_obj, outdir)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{Output directory}
}
\description{
Export TSRs as bed files
}
