% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_analysis.R
\name{export_TSSs}
\alias{export_TSSs}
\title{Export TSSs}
\usage{
export_TSSs(go_obj, outdir)
}
\arguments{
\item{go_obj}{gostripes object}

\item{outdir}{Output directory}
}
\description{
Export TSSs as bedgraphs
}
