% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_processing.R
\name{filter_soft}
\alias{filter_soft}
\title{Filter Soft-clipped}
\usage{
filter_soft(flagfiltered_bam, sample_name, outdir)
}
\arguments{
\item{flagfiltered_bam}{Bam file with PCR duplicates and undesirable flags filtered out}

\item{sample_name}{Name of the sample}

\item{outdir}{Output directory}
}
\description{
Filter reads with too many soft-clipped bases
}
