% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alignment.R
\name{genome_index}
\alias{genome_index}
\title{Generate Genome Index}
\usage{
genome_index(go_obj, genome_assembly, genome_annotation, outdir,
  cores = 1)
}
\arguments{
\item{go_obj}{gostripes object}

\item{genome_assembly}{Genome assembly in fasta file format}

\item{genome_annotation}{Genome annotation in gtf file format}

\item{outdir}{Directory to save genome index to}

\item{cores}{Number of CPU cores/threads to use}
}
\description{
Generate a STAR genomic index
}
