% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gostripes.R
\docType{class}
\name{gostripes-class}
\alias{gostripes-class}
\alias{gostripes}
\title{gostripes class}
\usage{
gostripes(sample_sheet, cores = 1)
}
\arguments{
\item{sample_sheet}{Sample sheet data.frame containing 'sample_name', 'replicate_ID", 'R1_read', and 'R2_read'}

\item{cores}{Number of CPU cores/threads available}
}
\description{
Class container appropriate slots for gostripes

Create gostripes object
}
\section{Slots}{

\describe{
\item{\code{sample_sheet}}{Sample sheet}

\item{\code{settings}}{Settings used throughout gostripes}
}}

