% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gostripes.R
\docType{class}
\name{gostripes-class}
\alias{gostripes-class}
\alias{gostripes}
\title{gostripes class}
\usage{
gostripes(sample_sheet, cores = 1)
}
\arguments{
\item{sample_sheet}{Sample sheet data.frame containing 'sample_name', 'replicate_ID", 'R1_read', and 'R2_read'}

\item{cores}{Number of CPU cores/threads available}
}
\value{
gostripes object containing the sample sheet and available cores
}
\description{
Class container appropriate slots for gostripes

Create gostripes object
}
\section{Slots}{

\describe{
\item{\code{sample_sheet}}{Sample sheet}

\item{\code{settings}}{Settings used throughout gostripes}

\item{\code{TSSs}}{List of GRanges containing transcription start sites}

\item{\code{TSRs}}{List of GRagnes containing transcription start regions (TSRs) or clustered transcription start sites (cTSSs)}

\item{\code{feature_counts}}{data.frame containing the RNA-seq like feature counting of reads}
}}

\examples{
R1_fastq <- system.file("extdata", "S288C_R1.fastq", package = "gostripes")
R2_fastq <- system.file("extdata", "S288C_R2.fastq", package = "gostripes")

sample_sheet <- tibble::tibble(
  "sample_name" = "stripeseq", "replicate_ID" = 1,
  "R1_read" = R1_fastq, "R2_read" = R2_fastq
)

go_object <- gostripes(sample_sheet)

}
